refactor(Car): 시도 횟수 입력 타입을 long으로 변경

+++ src/main/java/racingcar/dto/UserInputDTO.java
@@ -1,5 +1,5 @@
 package racingcar.dto;
 
-public record UserInputDTO(String names, int numOfStages) {
+public record UserInputDTO(String names, long numOfStages) {
 
 }
